add: Final logic
Switched from using a thread to only using timer tasks. The tasks for all ring times will be set at a specific time.
Added a trigger to HolidayController methods post, put, and delete to restart the timer tasks.
Added loadRingtimeTime to the speaker_api_settings.json file, where you can set the time when the ring times for the next day are loaded.