update: python sonosapi
Add method for checking if the speakers are in a group already.
Add discover() method to choose between the discover_by_name and discover_all method. Discover_by_name method is only used now, if some speakers are defined in the speaker_config.json file.
Update that the discovery methods return a sonos_speaker device instead of a list of sonos_speaker devices.