add: Argon2 Password Encoder Test
In this commit, we have added a new test case `testRawPasswordToHashedPassword` in `AuthenticationControllerTest.java` class. This test is used to validate the Argon2 password encoder functionality. It checks whether a raw password gets correctly hashed by the Argon2 password encoder and that the matches function can correctly verify this transformation. The tested password in this case is "Schulgong".

SCHUL-231