add: AuthResponseDTO and LoginDTO
In this commit, we have added two new Data Transfer Objects (DTOs): `AuthResponseDTO` and `LoginDTO`.

The `AuthResponseDTO` is used to encapsulate the authentication token in the response after a successful login operation. It contains a single field, `token`.

The `LoginDTO` is used to receive the login request from the client side. It contains a single field, `password`, which is the input password from the user. Both DTOs are equipped with Lombok annotations for automatic getter, setter, no-args constructor, all-args constructor, and builder pattern generation.

SCHUL-231