feature: Add SPA configuration for resource handling
This commit adds a new `SpaConfiguration` class which implements `WebMvcConfigurer`. This class is used to customize the default configurations of Spring MVC to properly handle static resources for a Single Page Application (SPA).

In the `addResourceHandlers` method, a handler is registered that serves static resources from the `/public/` directory located in the classpath for all paths.

The `PathResourceResolver` is overridden to return the requested resource if it exists and is readable. If the requested resource doesn't exist, it falls back to serving the `index.html` file located in the `/public/` directory. This is crucial for SPAs where most routes are client-side routes and do not have a corresponding server-side file or endpoint.

SCHUL-256